work on comments, refactoring
