+++ application.yaml
@@ -12,9 +12,6 @@ cmd/eskimo:
     port: 443
     certPath: cmd/eskimo/.testdata/localhost.crt
     keyPath: cmd/eskimo/.testdata/localhost.key
-  defaultPagination:
-    limit: 20
-    maxLimit: 1000
 cmd/eskimo-hut:
   host: localhost
   version: local
@@ -23,9 +20,6 @@ cmd/eskimo-hut:
     port: 443
     certPath: cmd/eskimo-hut/.testdata/localhost.crt
     keyPath: cmd/eskimo-hut/.testdata/localhost.key
-  defaultPagination:
-    limit: 20
-    maxLimit: 1000
 users:
   db:
     spaces:
@@ -37,7 +31,7 @@ users:
   messageBroker:
     #It's not used, because we`re not consuming anything in the production code.
     #So, its configured here just for testing or other local purposes.
-    consumerGroup: eskimo-users
+    consumerGroup: eskimo-users-testing
     createTopics: true
     urls:
       - localhost:9092
@@ -47,9 +41,6 @@ users:
         replicationFactor: 1
         retention: 1000h
   storage:
-    urlUpload: https://storage.bunnycdn.com
-    urlDownload: https://ice-staging.b-cdn.net
-    zoneName: ice-staging
-    profilePath: profile
+    urlUpload: https://storage.bunnycdn.com/ice-staging/profile
+    urlDownload: https://ice-staging.b-cdn.net/profile
     accessKey: access_key_hidden
-    retryCount: 3
